moved AstFile to `astFile.ts` and introduced new type so that `Tree` does not have to be imported anywhere except where explicitly needed
